% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_runtime.r
\name{comm_manager}
\alias{comm_manager}
\title{Get global CommManager instance}
\usage{
comm_manager()
}
\value{
\link{CommManager} instance if a kernel is running, else NULL
}
\description{
Get global CommManager instance
}
