% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernel.r
\name{installspec}
\alias{installspec}
\title{Install the kernelspec to tell IPython (>= 3) about IRkernel}
\usage{
installspec(user = T)
}
\arguments{
\item{user}{Install into user directory ~/.ipython or globally?}
}
\description{
Install the kernelspec to tell IPython (>= 3) about IRkernel
}

