% Generated by roxygen2 (4.0.0): do not edit by hand

\docType{package}
\name{IRkernel}
\alias{IRkernel}
\alias{IRkernel-package}
\title{An R kernel for IPython.}
\description{
IPython's modern interfaces, including the IPython
Notebook, speak a JSON+ZMQ protocol to a 'kernel' which is
responsible for executing code. This protocol is language
agnostic, so other languages can take advantage of
IPython's rich UI by implementing a kernel. This package is
a kernel for the R language.
}

