% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.r
\docType{class}
\name{Kernel-class}
\alias{Kernel}
\alias{Kernel-class}
\title{The kernel}
\description{
Has methods able to connect and talk to a Jupyter server.
}
\section{Methods}{

\describe{
\item{\code{abort_queued_messages()}}{Abort all already queued shell messages after an error}

\item{\code{abort_shell_msg()}}{Send an abort message for an incoming shell request}

\item{\code{handle_shell()}}{React to a shell message coming in}

\item{\code{is_complete(request)}}{Checks whether the code in the rest is complete}

\item{\code{msg_to_wire(msg)}}{Serialize a message}

\item{\code{new_reply(msg_type, parent_msg)}}{Prepare a reply}

\item{\code{send_response(msg_type, parent_msg, socket_name, content)}}{Send a response}

\item{\code{sign_msg(msg_lst)}}{Sign messages}

\item{\code{wire_to_msg(parts)}}{Deserialize a message}
}}

