% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{log_*}
\alias{log_*}
\alias{log_debug}
\alias{log_info}
\alias{log_error}
\title{Kernel logging functions}
\usage{
log_debug(...)

log_info(...)

log_error(...)
}
\arguments{
\item{...}{message to log}
}
\description{
A set of exported logging utilities that have the capability to be used in upstream projects.
Log level and log file can be set via R package options e.g. \code{options(jupyter.log_level = 2L)}
or from the environment variables JUPYTER_LOG_LEVEL and JUPYTER_LOGFILE.
}
