% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/execution.r
\name{set_plot_options}
\alias{set_plot_options}
\title{Set options for plotting}
\usage{
set_plot_options(...)
}
\arguments{
\item{...}{options that will be passed to  png()}
}
\description{
IRkernel displays plots in the notebook with calls to png().
This function allows to set the variables that will be passed on to
png(), for example width or height, see help(png).
}

