% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installspec.r
\name{installspec}
\alias{installspec}
\title{Install the kernelspec to tell Jupyter about IRkernel.}
\usage{
installspec(
  user = NULL,
  name = "ir",
  displayname = "R",
  rprofile = NULL,
  prefix = NULL,
  sys_prefix = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{user}{Install into user directory (\href{https://specifications.freedesktop.org/basedir-spec/latest/ar01s03.html}{\code{$XDG_DATA_HOME}}\code{/jupyter/kernels}) or globally? (default: NULL but treated as TRUE if "prefix" is not specified)}

\item{name}{The name of the kernel (default "ir")}

\item{displayname}{The name which is displayed in the notebook (default: "R")}

\item{rprofile}{(optional) Path to kernel-specific Rprofile (defaults to system-level settings)}

\item{prefix}{(optional) Path to alternate directory to install kernelspec into (default: NULL)}

\item{sys_prefix}{(optional) Install kernelspec using the \code{--sys-prefix} option of the currently detected jupyter (default: NULL)}

\item{verbose}{(optional) If \code{FALSE}, silence output of \code{install}}
}
\value{
Exit code of the \code{jupyter kernelspec install} call.
}
\description{
This can be called multiple times for different R interpreter, but you have to give a
different name (and displayname to see a difference in the notebook UI). If the same
name is give, it will overwrite older versions of the kernel spec with that name!
}
