% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{main}
\alias{main}
\title{Initialise and run the kernel}
\usage{
main(connection_file = "")
}
\arguments{
\item{connection_file}{The path to the Jupyter connection file, written by the frontend}
}
\description{
Initialise and run the kernel
}
